requirement.txt
asgiref==3.5.2
distlib==0.3.5
Django==4.1
django-braces==1.15.0
django-ckeditor==6.5.0
django-comments-dab==2.8.0
django-crispy-forms==1.14.0
django-filter==22.1
django-gravatar2==1.4.4
django-js-asset==2.0.0
django-model-utils==4.2.0
django-star-ratings==0.9.2
django-widget-tweaks==1.4.12
djangorestframework==3.13.1
env==0.1.0
ez-setup==0.9
filelock==3.6.0
Markdown==3.4.1
Pillow==9.1.0
platformdirs==2.5.2
PyJWT==1.7.1
python-decouple==3.6
pytz==2022.2.1
six==1.16.0
sqlparse==0.4.2
swapper==1.3.0
tzdata==2022.1
virtualenv==20.16.3
